WAITM-33: Deceased status in patient list table
PR: #1762

Included internal API change: table row styles are now generically computable.

Previously: special coupling where Table Rows looked for a `discontinued` field on the
row data, and would apply a special "striked off" look to that row. This was added for the
medication table work in TAN-904.

Now: Table gets a cool bean `rowStyle` prop which is a function that gets given the row
data and returns a style string. The medication table was updated to use this API, and
this PR was able to trivially style the rows without adding another tightly-coupled prop.