WAITM-12: VDS cryptography and translation (#1748)
This is a joining of two parts of VDS:
- implementing the cryptography of signing VDS-NC documents, as well as that for the
  generating and securing of key material for the VDS document signer.
- translating the Tamanu models and data into what is expected for the VDS documents,
  for both the Proof of Vaccination and Proof of Test document schemas.

As generating a VDS document involves signing it, and as the data structures involved
are rather entwined, this turned into a single unit of work rather than being split. Ref the
tickets WAITM-12 and WAITM-19.

As part of the test suite, the cryptography is tested end-to-end, and this requires having
a CSCA. Thus, this commit also includes a test implementation of generating a CSCA
keypair and certificate. This could be used for generating a longer-lived keypair for dev
and staging with minor adaptation, though it's not set up this way at the moment.

This commit also increases the minimum version of Node.js to be used to 12.16.0, as a
requirement for the use of ECDSA cryptography. This is not expected to disrupt anything:
indeed 12.16 is already in use as far as I could tell; this is a "just in case" enforcement of
the requirement.

A lot of the cryptography and document structure requires referring to various parts of
the ICAO 9303 set of documents. An important context note: the main guiding document
is the Technical Report which defines the VDS-NC standard. Some details in 9303 do
not make sense otherwise, especially around the certificate profile and field requirements
for Barcode Signers, which the VDS Signer technically is. Always refer to the TR first and
foremost, especially in situations where there appears to be a conflict between the code
here and what is specified in 9303.

This commit adds several scheduled tasks to support the signing and signer activity:
- one to sign pending Documents. This is discussed later.
- two to renew the Signer. One does the actual renewal crypto, and the other sends the
  renewal CSR email. This is so the email sending is automatically retried if it fails e.g.
  due to connectivity.
- one to erase a Signer's private key once the Signer has expired. This is an ICAO
  requirement.

Creating and signing documents is expected to be done in one step, but the design
allows for documents to be generated unsigned, for signing later. This is where the task
comes in. The expected uses of this alternate mechanism are:
- in a distant future where cryptography may be more physically separated, a more
  independent service could provide this with access to secrets, while the rest of the
  server would remain within the sync-server monolith.
- much more likely: if a Signer is not available, but one is expected to be available soon,
  documents could be generated in advance and be essentially "queued" for signing as
  soon as possible.