Always set updated_at_sync_tick, throwing away client version
This is because atm sync has the fundamental flaw that once a facility's sync cursor moves on, it will never retrieve changes sent in from other facilities that were updated before that cursor but synced in after it. It comes at the cost of last write wins conflict resolution, but this is a worthwhile trade, and we can solve that in another way later.