EPI-227: Fix dob/dod handling
In a previous revision I had used date formatting to days precision for these
fields, even though the fields are already in the correct format, to make date
use "consistent" -- however that introduces timezone issues due to the source
dates not having timezone information. This *could* be fixed differently by
reading the countryTimeZone setting instead but when the source data is quite
correct on its own there is really no point transforming it to itself.