Replace .push(...values) with concat
The former has a limitation in that large enough lengths of values will
throw a call stack error. This is unlikely to trigger at current scale
and for almost all affected workloads, but may be the case for imports,
reports, and sync, at large enough installations. Avoiding the pattern
altogether sidesteps this issue.