Fix tests-pass job
On advice from Github support, the dummy job needs to be set to always
run, so it can then evaluate if it should fail the build. Otherwise a
skipped or cancelled dependency, such as happens when the build is
cancelled in favour of a new commit on the PR, would cause the
tests-pass job to pass the build unwantedly.

Incidentally this construction also lets us be more flexible with the
fail condition if we wish to: for example, we could have the storybook
job run even though it's broken and specifically allow it through.