Fix docker build caching
Before the last of (central, facility) to finish in the previous run would partially overwrite the cache for both, and the other would therefore start the next run with a half-cold cache. The fix has a side-effect where the first build on any given branch will start from a cold cache.