Containerisation fixes (#4178)
* Fixes when running on a branch

* Prepare for linux-arm64 support

* Provision failed when some fields were missing

* Clarity

* Add labels

* Prefix commit hash with sha- so it's findable

* Pathing is hard

* Provide name/code from id if missing

* More nulls

* Allow arbitrary fields for users provision

* Nearly there

* Fix docker build caching

Before the last of (central, facility) to finish in the previous run would partially overwrite the cache for both, and the other would therefore start the next run with a half-cold cache. The fix has a side-effect where the first build on any given branch will start from a cold cache.