ci: fix release workflow
A committish in git parlance can be a branch, sha, tag, or [a number of other things][1]. However, it turns out that in the github REST API, it can only be a branch or sha. We used to use the tag here, but now switched to the exact sha ref.

[1]: https://stackoverflow.com/questions/23303549/what-are-commit-ish-and-tree-ish-in-git