ci(changelog): Clean up release note output
Cliff was outputting way too much, and not the correct things either, so
fix that by checking out the exact tag we want and instructing it to
write the "--current" release notes.

However, in this mode Cliff sometimes outputs more than one version's
worth of release notes, and there's no way to limit that behaviour, so
post-process the release notes to keep the top section only.

While we're at it, remove "no-issue" and "release-fix-vX.Y" labels
often present in PR titles but with zero value in the changelog.