feat(invoicing): EPI-846: backend for patient payment (#6057)
* feat: Add patient payment route and schema

This commit adds a new route and schema for creating patient payments. It also includes the necessary dependencies and helper functions.

* feat: Refactor patient payment schema and models

This commit refactors the patient payment schema and models to improve code organization and readability. It also updates the dependencies and includes helper functions for creating patient payments.

* feat: Refactor invoice payment schema and models

This commit refactors the invoice payment schema and models to improve code organization and readability. It also includes the necessary dependencies and updates the invoice model to include a patient payment status field.

* refactor: Rename 'method' field to 'method_id' in addInvoicePaymentsTable migration

* refactor: Remove constraints from InvoicePatientPayment and InvoicePayment associations

* fix: test error

* refactor: Update invoice item filtering logic in getInvoiceSummary function

* refactor: Update parameter name from 'id' to 'invoiceId' in patientPaymentRoute endpoints

* fix: accounting for the quantity of each invoice item

* refactor: Add product code to invoice items and update invoice summary logic

* Fix formatting and logic issues in code

---------

Co-authored-by: Huy-Tran-Huu-Le <160376020+Huy-Tran-Huu-Le@users.noreply.github.com>
Co-authored-by: khoi.nguyen <khoi@bes.au>