feat(pm2): opt into multi-process facility deployment (#6359)
* feat(pm2): simplify cwd to a single line

This makes it easier to use a fully-custom pm2 config by copying it outside of the release folder and adjusting the cwd line to point to the right place on upgrade.

* fix(pm2): use the canonical subcommand for startApi

* feat(pm2): use shorter pm2 process names

Less typing and better clarity in tables for $ pm2 status.

* feat(pm2): use fork mode for non-scaled processes

This has less overhead.

* feat(pm2): use memory/scale calculation for facility server

Like #6320

* feat(servers): always respect PORT env var

* tweak(servers): default sync api to port 5000

This is because port 400x is used for scaling out the api in a non-cluster load balancer.

* feat(pm2): run servers in fork mode

This requires external load-balancing (caddy) but reduces overhead, enables multiple different servers, and allows greater configuration opportunities.

* feat(pm2): default facility-server to multiprocess deployment

* tweak(fhir): run 2 separate fhir worker processes for refresh and resolver jobs

---------

Co-authored-by: Rohan Port <rohan@bes.au>