repo: drop the CSCA code
This does not drop the Signer and VDS/EUDCC code from Tamanu,
just the CSCA management bits. We no longer have any deployment
using this feature actively (no requests over the last 3+ months),
so we can hope that eventually we will remove the Tamanu code,
but it would be harder to revert so let's hold off on that a bit.