feat: EPI-1124: Ability to manually enter the price of an additional invoice product (#7298)
* feat: EPI-1124: Ability to manually enter the price of an additional invoice product

* add useEffect dependency and implement PriceField component

* update dbt model docs

* feat: Update price field validation to make it optional

* feat: Update price field validation for InvoiceProduct

* fix: Update visibility logic for price input in InvoiceItem component

This commit modifies the visibility logic for the price input field in the `InvoiceItemRow` component. The price input will now also be hidden when the item is not editable, enhancing the user experience by preventing unnecessary input fields from appearing in the UI.

* feat: Update price field validation and UI for InvoiceItem

The code changes in this commit update the price field validation and user interface for the `InvoiceItem` component. It adds support for optional price input and improves the display of crossed-out prices. The changes also include importing the `PriceField` component and the `REFERENCE_TYPES` constant from the `@tamanu/constants` module.

This commit addresses the following recent user commits:
- feat: Update price field validation for InvoiceProduct
- feat: Update price field validation to make it optional
- feat: EPI-1124: Ability to manually enter the price of an additional invoice product

It also takes into account the established commit message conventions in the repository.

* fix incorrect validation error message

* fix lint and display issue

* remove actions column in Patient payments