feat: Update price field validation and UI for InvoiceItem
The code changes in this commit update the price field validation and user interface for the `InvoiceItem` component. It adds support for optional price input and improves the display of crossed-out prices. The changes also include importing the `PriceField` component and the `REFERENCE_TYPES` constant from the `@tamanu/constants` module.

This commit addresses the following recent user commits:
- feat: Update price field validation for InvoiceProduct
- feat: Update price field validation to make it optional
- feat: EPI-1124: Ability to manually enter the price of an additional invoice product

It also takes into account the established commit message conventions in the repository.