feat: embed node engine in our dependencies
repo: change from .nvmrc to generic .node-version
ci: check that node versions specifications are all consistent

Because npm run's PATH includes all the relevant node_modules/.bin dirs
this change means that we always use the exact correct Node.js version.

It also means that we no longer need to install Node.js on Windows when
deploying servers, and that updating the nodejs version here will update
it on Windows deployments, too. That means we can opt into newer Node.js
much faster, without requiring extremely long runways for the deployment
team to coordinate runtime upgrades.

Further, this commit changes from the legacy .nvmrc file to the common
standard .node-version as described, with a symlink to keep nvm working:
<https://github.com/shadowspawn/node-version-usage>

Finally, we add a CI check to make sure that the three different places
that the node version is specified are kept in sync.