test(e2e): Add tests for outpatient and inpatient page (#8455)
* Adding tests for inpatients page search

* Applying AI comments and some fixes on all patients tests

* Adding more tests to inpatient test file

Adding clear search test
adding inpatient table pagination tests
adding sorting inpatient table tests

* Adding tests for outpatient page

* Refactoring patients pages to have a base class as BasePatientListPage

applying some AI comments and refactoring patients page code to have a base class as BasePatientListPage to remove the duplicate code.

* Changing the encounter location to a facility1 location instead of facility 2

Because the e2e tests are running on facility 1 in CI/CD we need to change the encounters created to be facility 1 instead of facility 2

* Adding the missing test data fields

* skipping a few tests due to the bug exist in search by department in the app

* skipping this test for now as it is failing in ci because of less than 10 outpatients and inpatients

* skipping this test for now as it is failing in ci because of less than 10 inpatients

* Changing the departmentId to cardiology for facility 1

The departmentId was referring to department in facility 2 and was causing the search by department tests failing, I am correcting that issue and removing the skip test.

* removing the skip test in out patient

* Refactoring the code to avoid repeating page.getByTestId so many times.

Refactoring the code to avoid repeating page.getByTestId so many times.